Refactor: Rename ProviderViewModel to ProviderDetailViewModel and RatesScreen path.
The `ProviderViewModel` has been renamed to `ProviderDetailViewModel` for clarity, and the `RatesScreen` has been moved from `com.currency.rateman.ui.screens` to `com.currency.rateman.provider.ui.screen`.

This change updates the view model name and file path to better reflect its purpose and improve code organization.
