+++ app/src/main/java/com/currency/rateman/di/KoinModules.kt
@@ -7,7 +7,7 @@ import com.currency.rateman.provider.data.repository.ProviderRepositoryImpl
 import com.currency.rateman.data.repository.SettingsRepository
 import com.currency.rateman.data.repository.SettingsRepositoryImpl
 import com.currency.rateman.ui.viewmodels.CurrencyViewModel
-import com.currency.rateman.provider.ui.viewmodel.ProviderViewModel
+import com.currency.rateman.provider.ui.viewmodel.ProviderDetailViewModel
 import com.currency.rateman.ui.viewmodels.RatesViewModel
 import com.currency.rateman.ui.viewmodels.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -26,5 +26,5 @@ val appModule = module {
     viewModel { RatesViewModel(get(), get()) }
     viewModel { SettingsViewModel(get()) }
     viewModel { CurrencyViewModel() }
-    viewModel { ProviderViewModel(get()) }
+    viewModel { ProviderDetailViewModel(get()) }
 }
